feat: adjust routing
